requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
bcrypt